\name{square}
\alias{square}

\title{square}

\description{Function to calculate the square of a number.}

\usage{
square(x)
}

\arguments{
  \item{x}{A numeric vector.}
}

\value{The value returned is a vector. Its element/s is/are of class numeric.}

\author{Alessandro Feller <ale_feller@hotmail.com>}

\examples{
x <- c(0:10,5)
square(x)
}